requirements.txt
fastapi
qdrant-client
redis
python-dotenv
lz4
orjson
asyncio
uvicorn